Extract remote address parser into dedicated file
For easier re-use in the upcoming pg-tunneling logic
