+++ server/src/main/java/org/elasticsearch/transport/RemoteConnectionStrategy.java
@@ -23,9 +23,6 @@ import static org.elasticsearch.common.settings.Setting.boolSetting;
 import static org.elasticsearch.common.settings.Setting.timeSetting;
 
 import java.io.Closeable;
-import java.net.InetAddress;
-import java.net.InetSocketAddress;
-import java.net.UnknownHostException;
 import java.util.Iterator;
 import java.util.Locale;
 import java.util.Map;
@@ -162,42 +159,6 @@ public abstract class RemoteConnectionStrategy implements TransportConnectionLis
             default:
                 throw new AssertionError("Invalid connection strategy" + mode);
         }
-    }
-
-    static InetSocketAddress parseConfiguredAddress(String configuredAddress) {
-        final String host = parseHost(configuredAddress);
-        final int port = parsePort(configuredAddress);
-        InetAddress hostAddress;
-        try {
-            hostAddress = InetAddress.getByName(host);
-        } catch (UnknownHostException e) {
-            throw new IllegalArgumentException("unknown host [" + host + "]", e);
-        }
-        return new InetSocketAddress(hostAddress, port);
-    }
-
-    static String parseHost(final String configuredAddress) {
-        return configuredAddress.substring(0, indexOfPortSeparator(configuredAddress));
-    }
-
-    static int parsePort(String remoteHost) {
-        try {
-            int port = Integer.parseInt(remoteHost.substring(indexOfPortSeparator(remoteHost) + 1));
-            if (port <= 0) {
-                throw new IllegalArgumentException("port number must be > 0 but was: [" + port + "]");
-            }
-            return port;
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException("failed to parse port", e);
-        }
-    }
-
-    private static int indexOfPortSeparator(String remoteHost) {
-        int portSeparator = remoteHost.lastIndexOf(':'); // in case we have a IPv6 address ie. [::1]:9300
-        if (portSeparator == -1 || portSeparator == remoteHost.length()) {
-            throw new IllegalArgumentException("remote hosts need to be configured as [host:port], found [" + remoteHost + "] instead");
-        }
-        return portSeparator;
     }
 
     /**
